refactor(composition): consolidate factory modules from 14 to 8 files
Reduce code duplication and improve navigation by consolidating factory modules:

- pipeline_factory.py: merged generic_factory + runner_assembly + runner_services
- services_factory.py: merged base_services_factory + services_builder
- data_source_factory.py: merged data_sources + data_source_registry
- storage.py: merged storage_factory + storage_adapter

Deleted obsolete files:
- clients.py (empty backward compat)
- generic_factory.py, runner_assembly.py, runner_services.py
- base_services_factory.py, services_builder.py
- data_sources.py, data_source_registry.py
- storage_factory.py, storage_adapter.py

Updated all imports across the codebase to use new consolidated modules.
All 135 relevant tests pass.
