--- src/bioetl/composition/factories/clients.py
@@ -1,12 +1,0 @@
-"""Factories for creating infrastructure clients.
-
-This module is simplified for local-only deployment.
-Previous cloud clients (Redis, S3) have been removed.
-"""
-
-from __future__ import annotations
-
-# This module is kept for backward compatibility
-# All cloud client factories have been removed for local-only deployment
-
-__all__: list[str] = []
