+++ src/bioetl/composition/providers/registration.py
@@ -49,7 +49,7 @@ def _get_factories() -> tuple[type, type]:
     Returns:
         Tuple of (DataSourceFactory, HttpClientFactory)
     """
-    from bioetl.composition.factories.data_sources import DataSourceFactory
+    from bioetl.composition.factories.data_source_factory import DataSourceFactory
     from bioetl.composition.factories.http_client_factory import HttpClientFactory
 
     return DataSourceFactory, HttpClientFactory
